update load_util.py
now the number of channels is dim 0, and the getitem function of dataset return torch.tensor instead of ndarray now

+++ src/file_utils/load_util.py
@@ -38,8 +38,8 @@ def load_images_masks(list_of_data_dir, list_of_mask_dir):
             a_x = np.swapaxes(a_x, 0, 1)
             a_y = np.swapaxes(a_y, 0, 1)
             print(f"after swap: {a_x.shape}")
-        a_x = cv2.resize(a_x, (2480, 3508)).astype(np.int16)
-        a_y = cv2.resize(a_y, (2480, 3508)).astype(np.int16)
+        a_x = cv2.resize(a_x, (2480, 3508)).astype(np.float32)
+        a_y = cv2.resize(a_y, (2480, 3508)).astype(np.float32)
         print(f"after resize: {a_x.shape}")
         x_array = xr.DataArray(
         data=a_x,
